🐛(aws_sign) add port to host header if present in uri
The Host header is set using the request uri to sign. If this host as a port,
we must also set it in the Host header. The uri.host() method doesn't
return the port if present. We have to test the presence of a port using
the uri.port() method and include it if not none.