🔧(action) dedicated action to docker build and publish
We add a github action to test docker image building. This action also
uses Trivy to scan the built image to detect CVE inside it and it is not
possible to publish if the scan detects CVEs. The publish is made on the
push on the main branch and whtn a tag v.* is created.